1. Create in toolbar_menus WebSocket. 2. Get data from WebSocket 3. Save new chatMsg from WebSocket

+++ app/src/main/java/ua/com/merchik/merchik/data/WebSocketData/WSChat.java
@@ -30,7 +30,7 @@ public class WSChat {
 
     @SerializedName("chat_id")
     @Expose
-    public Long chatId;
+    public Integer chatId;
 
     @SerializedName("dt_read")
     @Expose
